Move tests to src/test/suite
