+++ TODO
@@ -34,8 +34,6 @@ Improvements:
                 ☐ deleteFile function
                 ☐ writeToFile function
             Preferences API:
-                ✔ Get Main Kt @done(19-01-29 15:22)
-                ✔ Set Main Kt @done(19-02-06 15:06)
                 ✔ Get WPILib Version @done(19-02-06 15:06)
                 ✔ Set WPILib Version @done(19-02-06 15:06)
                 ✔ Get Compliance Tests Value @done(19-02-06 15:06)
