+++ src/test/suite/template_interpreter.test.ts
@@ -5,24 +5,24 @@ import * as assert from 'assert';
 // import * as commands from '../commands';
 // import * as file_generator from '../file_generator';
 // import * as preferences from '../preferences';
-import * as template_interpreter from '../templates/template_interpreter';
+import * as template_interpreter from '../../templates/template_interpreter';
 // import * as path from "path";
 // import * as fs from 'fs';
 import * as testingConsts from "./testingConstants";
-import * as consts from "../constants";
-import { MainTemplate } from '../templates/frc-kotlin/Main';
-import { BuildGradleTemplate } from '../templates/frc-kotlin/BuildGradle';
-import { OldCommandSubsystemTemplate } from '../templates/frc-kotlin/old-command-based/subsystems/SubsystemTemplate';
-import { OldCommandTemplate } from '../templates/frc-kotlin/old-command-based/commands/CommandTemplate';
-import { OldCommandGroupTemplate } from '../templates/frc-kotlin/old-command-based/commands/CommandGroupTemplate';
-import { EmptyClassTemplate } from '../templates/frc-kotlin/EmptyClassTemplate';
-import { OldCommandInstantCommandTemplate } from '../templates/frc-kotlin/old-command-based/commands/InstantCommandTemplate';
-import { OldCommandTimedCommandTemplate } from '../templates/frc-kotlin/old-command-based/commands/TimedCommand';
-import { OldCommandPIDSubsystemTemplate } from '../templates/frc-kotlin/old-command-based/subsystems/PIDSubsystemTemplate';
-import { OldCommandTriggerTemplate } from '../templates/frc-kotlin/old-command-based/triggers/TriggerTemplate';
-import { OldCommandRobotTemplate } from '../templates/frc-kotlin/old-command-based/Robot';
-import { OldCommandOITemplate } from '../templates/frc-kotlin/old-command-based/OI';
-import { OldCommandRobotMapTemplate } from '../templates/frc-kotlin/old-command-based/RobotMap';
+import * as consts from "../../constants";
+import { MainTemplate } from '../../templates/frc-kotlin/Main';
+import { BuildGradleTemplate } from '../../templates/frc-kotlin/BuildGradle';
+import { OldCommandSubsystemTemplate } from '../../templates/frc-kotlin/old-command-based/subsystems/SubsystemTemplate';
+import { OldCommandTemplate } from '../../templates/frc-kotlin/old-command-based/commands/CommandTemplate';
+import { OldCommandGroupTemplate } from '../../templates/frc-kotlin/old-command-based/commands/CommandGroupTemplate';
+import { EmptyClassTemplate } from '../../templates/frc-kotlin/EmptyClassTemplate';
+import { OldCommandInstantCommandTemplate } from '../../templates/frc-kotlin/old-command-based/commands/InstantCommandTemplate';
+import { OldCommandTimedCommandTemplate } from '../../templates/frc-kotlin/old-command-based/commands/TimedCommand';
+import { OldCommandPIDSubsystemTemplate } from '../../templates/frc-kotlin/old-command-based/subsystems/PIDSubsystemTemplate';
+import { OldCommandTriggerTemplate } from '../../templates/frc-kotlin/old-command-based/triggers/TriggerTemplate';
+import { OldCommandRobotTemplate } from '../../templates/frc-kotlin/old-command-based/Robot';
+import { OldCommandOITemplate } from '../../templates/frc-kotlin/old-command-based/OI';
+import { OldCommandRobotMapTemplate } from '../../templates/frc-kotlin/old-command-based/RobotMap';
 
 // TODO: Differentiate between old and new command based
 suite("Grabbing Templates", function () {
