+++ src/test/suite/testingConstants.ts
@@ -1,13 +1,16 @@
+import { resolve } from 'path';
 import * as spawn from "cross-spawn";
 import * as vscode from "vscode";
-import * as kotlinExt from "../extension";
+import * as kotlinExt from "../../extension";
 
 export function setupWorkspace() {
-    if (typeof vscode.workspace.workspaceFolders === 'undefined') {
-        console.log("Whoops");
-        return;
-    }
-    var uri = vscode.Uri.parse(vscode.workspace.workspaceFolders[0].uri.path + "/testing-workspace/workspace");
+    // if (typeof vscode.workspace.workspaceFolders === 'undefined') {
+    //     console.log("Whoops");
+    //     return;
+    // }
+    // console.log(process.cwd());
+    console.log(resolve(__dirname, "..", "..", ".."));
+    var uri = vscode.Uri.parse(resolve(__dirname, "..", "..", "..") + "/testing-workspace/workspace");
     kotlinExt.setWorkspaceFolderFsPath(uri.fsPath);
     kotlinExt.setWorkspaceFolderPath(uri.path);
 }
